requirements.txt
# ID-SEC: [JNL-MVP-NUVEM-ANALIZADOR]
pandas
openpyxl
python-docx